% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_get_modules.R
\name{get_modules}
\alias{get_modules}
\title{Get Modules}
\usage{
get_modules(selected_course, modules_init, acad_year)
}
\arguments{
\item{selected_course}{Selected course, can be "All" or specific.}

\item{modules_init}{The table from which it gets the modules.}

\item{acad_year}{The academic year of the users.}
}
\value{
A vector of module name by module and title
}
\description{
Get a vector of module named by module and title from a selected course
}
\examples{
# Need a reactive context and a MongoDB database
}
