% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_get_users.R
\name{get_users}
\alias{get_users}
\title{Get Users}
\usage{
get_users(selected_course, only_enrolled, sdd_users2)
}
\arguments{
\item{selected_course}{Selected course, can be "All" or specific}

\item{only_enrolled}{Only the enrolled users or everyone}

\item{sdd_users2}{The table from which we get the users}
}
\value{
A vector of users_id name by logins and lastnames
}
\description{
Get a vector of users_id named by logins and lastnames from a selected course and with enrolled only or everyone
}
\examples{
Need a reactive context and a MongoDB database
}
