% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_html_link_with_app_name.R
\name{html_link_with_app_name}
\alias{html_link_with_app_name}
\title{HTML Link with App Names}
\usage{
html_link_with_app_name(data_frame)
}
\arguments{
\item{data_frame}{data.frame from which it takes the url and app.}
}
\value{
A vector of html links.
}
\description{
HTML Link with App Names
}
\examples{
# Need special data frame
}
