% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_is_request.R
\name{is_request}
\alias{is_request}
\title{is_request function}
\usage{
is_request(input)
}
\arguments{
\item{input}{The input to check if there is a request}
}
\value{
TRUE if there is a request about this input, FALSE if not
}
\description{
is_request function
}
\examples{
# Need a reactive context
}
