% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnItDashboard-package.R
\docType{package}
\name{LearnItDashboard-package}
\alias{LearnItDashboard-package}
\title{LearnItDashboard}
\description{
This is a dashboard app to explore learnitdown data
}
\section{Application Shiny}{


- To launch the Shiny application : use [run_app()]
}

