% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_my_glue.R
\name{my_glue}
\alias{my_glue}
\title{My Glue}
\usage{
my_glue(...)
}
\arguments{
\item{...}{Expression to be glued with "<<" and ">>" delimiters.}
}
\value{
The result of the glue::glue function using "<<" and ">>" delemiters.
}
\description{
My Glue
}
\examples{
my_glue("2 + 2 = <<2+2>>")
}
