% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prepare_content.R
\name{prepare_content}
\alias{prepare_content}
\title{Prepare Content}
\usage{
prepare_content(data_frame)
}
\arguments{
\item{data_frame}{data.frame on which prepare the content}
}
\value{
A vector of html links.
}
\description{
Prepare a given dataframe's content with html tags, such as links or images, to be rendered in a timeline.
}
\examples{
# Need special data frame
}
