% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prepare_request.R
\name{prepare_request}
\alias{prepare_request}
\title{Pepare Request}
\usage{
prepare_request(request_vector, args, type = NULL)
}
\arguments{
\item{request_vector}{Vector of prepared arguments from selected choices.}

\item{args}{Vector of selected arguments to use in the request.}

\item{type}{Type of table to query. Can be h5p, learnr or shiny.}
}
\value{
A string request with the chosen arguments.
}
\description{
Prepare the request to get data from MongoDB from chosen arguments and a vector containing different ones.
}
\examples{
rv <- c("color" = "red", "type" = "flower")
prepare_request(rv, "color")
}
