% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prepare_style.R
\name{prepare_style}
\alias{prepare_style}
\title{Prepare Style}
\usage{
prepare_style(data_frame)
}
\arguments{
\item{data_frame}{data.frame on which prepare the content.}
}
\value{
A vector of css styles
}
\description{
Prepare a given data frame's css style, such as background-color, to be
rendered in a timeline.
}
\examples{
# Need special data frame
}
