% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prepare_url_alt_url.R
\name{prepare_url_alr_url}
\alias{prepare_url_alr_url}
\title{Prepare Url and Alt_Url}
\usage{
prepare_url_alr_url(data_table)
}
\arguments{
\item{data_table}{The data frame from which the url and alt_url are prepared.}
}
\value{
The given data frame with the url and alt_url prepared as html link.
}
\description{
Prepare the url and alt_url of a given dataframe, making them an html link.
}
\examples{
# Need a special data frame
}
