% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_progression_plot.R
\name{progression_plot}
\alias{progression_plot}
\title{Plot the Progression of the Students in the Data Science Exercises}
\usage{
progression_plot(login, course, module = NULL, sdd_url)
}
\arguments{
\item{login}{The student (GitHub) login.}

\item{course}{The institutional identifier of the course, e.g., "S-BIOG-015".}

\item{module}{The module to restrict to, or NULL for all modules in the course.}

\item{sdd_url}{The URL to the SDD MongoDB database.}
}
\value{
This function is used for its side-effect of plotting a chart comparing the student and the class progression.
}
\description{
Plot the Progression of the Students in the Data Science Exercises
}
\examples{
#progression_plot(<insert_login_here, "S-BIOG-015",
#  module = NULL, sdd_url = "mongodb://127.0.0.1:27017/sdd")
}
\author{
: Philippe Grosjean (phgrosjean\@sciviews.org)
}
