% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnitdashboard-package.R
\docType{package}
\name{learnitdashboard-package}
\alias{learnitdashboard-package}
\title{learnitdashboard}
\description{
This is a dashboard app to explore learnitdown data
}
\section{Application Shiny}{


- To launch the Shiny application : use [run_app()]
}

