% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnitdashboard-package.R
\docType{package}
\name{learnitdashboard-package}
\alias{learnitdashboard}
\alias{learnitdashboard-package}
\title{learnitdashboard: Shiny Dashboard for 'learnitr' Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A dashboard to explore data generated by 'learnitr' code, such as the 'learnitdown', or the 'learnitgrid' R package. This is a lightweight learning analytics solution for the 'learnitr' ecosystem.
}
\details{
This is a dashboard app to explore 'learnitr' data.
}
\section{Shiny application}{

\itemize{
\item To launch the Shiny application : use \code{\link[=run_app]{run_app()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/learnitr/learnitdashboard}
  \item \url{https://learnitr.github.io/learnitdashboard/}
  \item \url{https://learnitr.r-universe.dev/learnitdashboard}
  \item Report bugs at \url{https://github.com/learnitr/learnitdashboard/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

Authors:
\itemize{
  \item Martin Lecocq \email{martin.lecocq@std.heh.be}
}

}
\keyword{internal}
