% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_prepare_url_alt_url.R
\name{prepare_url_alr_url}
\alias{prepare_url_alr_url}
\title{Prepare Url and Alt_Url}
\usage{
prepare_url_alr_url(data_table)
}
\arguments{
\item{data_table}{The dataframe from which the url and alt_url have to be prepared.}
}
\value{
The given dataframe with the url and alt_url prepared in html link format.
}
\description{
Prepare the url and alt_url of a given dataframe, making them an html link.
}
\examples{
# Need a special dataframe
}
